+++ src/main/java/me/devflare/CraftGuard/CraftGuard.java
@@ -2,11 +2,14 @@ package me.devflare.CraftGuard;
 
 import me.devflare.CraftGuard.commands.CraftGuardCommand;
 import me.devflare.CraftGuard.config.ConfigManager;
+import me.devflare.CraftGuard.listeners.ContainerListener;
 import me.devflare.CraftGuard.listeners.CraftingListener;
 import me.devflare.CraftGuard.listeners.PortalListener;
 import me.devflare.CraftGuard.listeners.WorkstationListener;
 import me.devflare.CraftGuard.placeholders.CraftGuardExpansion;
+import me.devflare.CraftGuard.utils.AsyncAuditLogger;
 import me.devflare.CraftGuard.utils.MessageUtil;
+import me.devflare.CraftGuard.utils.WorldGuardHook;
 import org.bukkit.Bukkit;
 import org.bukkit.command.PluginCommand;
 import org.bukkit.event.HandlerList;
@@ -16,16 +19,26 @@ import org.bukkit.plugin.java.JavaPlugin;
  * CraftGuard - World-based crafting management for Minecraft servers
  * 
  * @author DevFlare, ItzzMateo
- * @version 1.2.2
+ * @version 1.4.0
  */
 public final class CraftGuard extends JavaPlugin {
 
     private static CraftGuard instance;
     private ConfigManager configManager;
+    private me.devflare.CraftGuard.utils.GUIManager guiManager;
+    private me.devflare.CraftGuard.listeners.GUIListener guiListener;
     private CraftGuardExpansion placeholderExpansion;
     private CraftingListener craftingListener;
     private PortalListener portalListener;
     private WorkstationListener workstationListener;
+    private ContainerListener containerListener;
+    private WorldGuardHook worldGuardHook;
+    private AsyncAuditLogger auditLogger;
+
+    @Override
+    public void onLoad() {
+        WorldGuardHook.registerFlag();
+    }
 
     @Override
     public void onEnable() {
@@ -44,8 +57,11 @@ public final class CraftGuard extends JavaPlugin {
         configManager = new ConfigManager(this);
         getLogger().info(configManager.getMessage("config-loaded"));
 
+        // Initialize GUI Manager
+        guiManager = new me.devflare.CraftGuard.utils.GUIManager(this);
+
         // Register command
-        CraftGuardCommand commandHandler = new CraftGuardCommand(this);
+        CraftGuardCommand commandHandler = new CraftGuardCommand(this, guiManager);
         PluginCommand command = getCommand("craftguard");
         if (command != null) {
             command.setExecutor(commandHandler);
@@ -54,6 +70,10 @@ public final class CraftGuard extends JavaPlugin {
         } else {
             getLogger().severe(configManager.getMessage("command-registration-failed"));
         }
+
+        // Initialize Hooks & Utilities
+        worldGuardHook = new WorldGuardHook();
+        auditLogger = new AsyncAuditLogger(this);
 
         // Register event listeners
         registerListeners();
@@ -91,6 +111,20 @@ public final class CraftGuard extends JavaPlugin {
             HandlerList.unregisterAll(workstationListener);
             workstationListener = null;
         }
+        if (containerListener != null) {
+            HandlerList.unregisterAll(containerListener);
+            containerListener = null;
+        }
+        if (guiListener != null) {
+            HandlerList.unregisterAll(guiListener);
+            guiListener = null;
+        }
+
+        // Shutdown Logger
+        if (auditLogger != null) {
+            auditLogger.shutdown();
+            auditLogger = null;
+        }
 
         // Unregister PlaceholderAPI expansion
         if (placeholderExpansion != null) {
@@ -120,10 +154,41 @@ public final class CraftGuard extends JavaPlugin {
         craftingListener = new CraftingListener(this);
         portalListener = new PortalListener(this);
         workstationListener = new WorkstationListener(this);
+        containerListener = new ContainerListener(this);
+        guiListener = new me.devflare.CraftGuard.listeners.GUIListener(this, guiManager);
 
         Bukkit.getPluginManager().registerEvents(craftingListener, this);
         Bukkit.getPluginManager().registerEvents(portalListener, this);
         Bukkit.getPluginManager().registerEvents(workstationListener, this);
+        Bukkit.getPluginManager().registerEvents(containerListener, this);
+        Bukkit.getPluginManager().registerEvents(guiListener, this);
+    }
+
+    /**
+     * Get the GUI manager
+     *
+     * @return GUIManager instance
+     */
+    public me.devflare.CraftGuard.utils.GUIManager getGuiManager() {
+        return guiManager;
+    }
+
+    /**
+     * Get the WorldGuard hook
+     *
+     * @return WorldGuardHook instance
+     */
+    public WorldGuardHook getWorldGuardHook() {
+        return worldGuardHook;
+    }
+
+    /**
+     * Get the audit logger
+     *
+     * @return AsyncAuditLogger instance
+     */
+    public AsyncAuditLogger getAuditLogger() {
+        return auditLogger;
     }
 
     /**